requirements.txt
ahrs==0.3.1
bagpy==0.4.6
deap==1.3.1
encode==0.0.0
flax==0.4.0
matplotlib==3.4.3
ml_collections==0.1.1
norse==0.0.7.post1
numpy==1.19.5
objgraph==3.5.0
optax==0.1.1
pandas==1.2.4
pympler==1.0.1
PyYAML==6.0
scipy==1.6.2
setuptools==52.0.0.post20210125
torchvision==0.11.3
tqdm==4.59.0
